Update README.md
The following is an examination of the design for a payment processing system. This design is guided by chapter 11 of the book System Design Interview: An Insider's Guide by Alex Xu and Sahn Lam. E-commerce has become quite prevalent in the wake of COVID-19 and a reliable payment system is essential to that operation. This system design is a look into a consistent, scalable, and flexible payment system.
The design begins with a simple, high-level design and continues deeper into the payment system components, including the initial HTML server/client connection, resources planning, the final implementation, and security considerations.